requirements.txt
pandas
pytest
openai
python-dotenv
sqlparse
